finops-ai-teams-bot/requirements.txt
flask
openai
azure-identity
azure-mgmt-resource
azure-mgmt-costmanagement
botbuilder-core
botbuilder-schema
pytest
